--- a/cron/hit_stat.go
+++ b/cron/hit_stat.go
@@ -6,21 +6,15 @@ package cron
 
 import (
 	"context"
-	"time"
+	"database/sql"
+	"strconv"
 
-	"github.com/jinzhu/now"
+	"github.com/pkg/errors"
 	"zgo.at/goatcounter"
 	"zgo.at/utils/jsonutil"
-	"zgo.at/utils/sliceutil"
 	"zgo.at/zdb"
 	"zgo.at/zdb/bulk"
 )
-
-type stat struct {
-	Path      string    `db:"path"`
-	Count     int       `db:"count"`
-	CreatedAt time.Time `db:"created_at"`
-}
 
 // Hit stats are stored per day/path, the value is a 2-tuple: it lists the
 // counts for every hour. The first is the hour, second the number of hits in
@@ -32,6 +26,9 @@ type stat struct {
 //   stats      | [[0,0],[1,2],[2,2],[3,0],[4,0],[5,0],[6,1],[7,2],[8,3],
 //                 [9,0],[10,2],[11,2],[12,2],[13,5],[14,4],[15,3],[16,0],
 //                 [17,1],[18,2],[19,0],[20,0],[21,1],[22,4],[23,2]]
+//
+// TODO: this can either just assume hour by index, or not store all the hours.
+// TODO: need to fill in blank days.
 func updateHitStats(ctx context.Context, phits map[string][]goatcounter.Hit) error {
 	txctx, tx, err := zdb.Begin(ctx)
 	if err != nil {
@@ -43,8 +40,7 @@ func updateHitStats(ctx context.Context, phits map[string][]goatcounter.Hit) err
 
 	// Group by day + path.
 	type gt struct {
-		// TODO: count should be [][]int
-		count int
+		count [][]int
 		day   string
 		path  string
 	}
@@ -54,42 +50,43 @@ func updateHitStats(ctx context.Context, phits map[string][]goatcounter.Hit) err
 			day := h.CreatedAt.Format("2006-01-02")
 			k := day + h.Path
 			v := grouped[k]
-			if v.count == 0 {
+			if len(v.count) == 0 {
 				v.day = day
 				v.path = h.Path
 
-				// TODO
 				// Append existing and delete from DB; this will be faster than
 				// running an update for every row.
-				// var c int
-				// err := tx.GetContext(txctx, &c, `select count from location_stats where site=$1 and day=$2 and location=$3`,
-				// 	h.Site, day, v.location)
-				// if err != sql.ErrNoRows {
-				// 	if err != nil {
-				// 		return errors.Wrap(err, "existing")
-				// 	}
-				// 	_, err = tx.ExecContext(txctx, `delete from location_stats where site=$1 and day=$2 and location=$3`,
-				// 		h.Site, day, v.location)
-				// 	if err != nil {
-				// 		return errors.Wrap(err, "delete")
-				// 	}
-				// }
+				var c []byte
+				err := tx.GetContext(txctx, &c,
+					`select stats from hit_stats where site=$1 and day=$2 and path=$3`,
+					h.Site, day, v.path)
+				if err != sql.ErrNoRows {
+					if err != nil {
+						return errors.Wrap(err, "existing")
+					}
+					_, err = tx.ExecContext(txctx,
+						`delete from hit_stats where site=$1 and day=$2 and path=$3`,
+						h.Site, day, v.path)
+					if err != nil {
+						return errors.Wrap(err, "delete")
+					}
+				}
 
-				//v.count = c
+				if c != nil {
+					jsonutil.MustUnmarshal(c, &v.count)
+				} else {
+					v.count = make([][]int, 24)
+					for i := range v.count {
+						v.count[i] = []int{i, 0}
+					}
+				}
 			}
 
-			v.count += 1
+			h, _ := strconv.ParseInt(h.CreatedAt.Format("15"), 10, 8)
+			v.count[h][1] += 1
 			grouped[k] = v
 		}
 	}
-
-	// TODO
-	// hourly := fillHitBlanks(stats, existing, site.CreatedAt)
-
-	// // No data received.
-	// if len(hourly) == 0 {
-	// 	return nil
-	// }
 
 	siteID := goatcounter.MustGetSite(ctx).ID
 	ins := bulk.NewInsert(ctx, zdb.MustGet(ctx),
@@ -103,96 +100,4 @@ func updateHitStats(ctx context.Context, phits map[string][]goatcounter.Hit) err
 	}
 
 	return tx.Commit()
-}
-
-func fillHitBlanks(stats []stat, existing []string, siteCreated time.Time) map[string]map[string][][]int {
-	// Convert data to easier structure:
-	// {
-	//   "jquery.html": map[string][][]int{
-	//     "2019-06-22": []{
-	// 	     []int{4, 50},
-	// 	     []int{6, 4},
-	// 	   },
-	// 	   "2019-06-23": []{ .. }.
-	// 	 },
-	// 	 "other.html": { .. },
-	// }
-	hourly := map[string]map[string][][]int{}
-	first := now.BeginningOfDay()
-	for _, s := range stats {
-		_, ok := hourly[s.Path]
-		if !ok {
-			hourly[s.Path] = map[string][][]int{}
-		}
-
-		if s.CreatedAt.Before(first) {
-			first = now.New(s.CreatedAt).BeginningOfDay()
-		}
-
-		day := s.CreatedAt.Format("2006-01-02")
-		hourly[s.Path][day] = append(hourly[s.Path][day],
-			[]int{s.CreatedAt.Hour(), s.Count})
-	}
-
-	// Fill in blank days.
-	n := now.BeginningOfDay()
-	alldays := []string{first.Format("2006-01-02")}
-	for first.Before(n) {
-		first = first.Add(24 * time.Hour)
-		alldays = append(alldays, first.Format("2006-01-02"))
-	}
-	allhours := make([][]int, 24)
-	for i := 0; i <= 23; i++ {
-		allhours[i] = []int{i, 0}
-	}
-	for path, days := range hourly {
-		for _, day := range alldays {
-			_, ok := days[day]
-			if !ok {
-				hourly[path][day] = allhours
-			}
-		}
-
-		// Backlog new paths since site start.
-		// TODO: would be better to modify display logic, instead of storing
-		// heaps of data we don't use.
-		if !sliceutil.InStringSlice(existing, path) {
-			ndays := int(time.Now().UTC().Sub(siteCreated) / time.Hour / 24)
-			daysSinceCreated := make([]string, ndays)
-			for i := 0; i < ndays; i++ {
-				daysSinceCreated[i] = siteCreated.Add(24 * time.Duration(i) * time.Hour).Format("2006-01-02")
-			}
-
-			for _, day := range daysSinceCreated {
-				if _, ok := hourly[path][day]; !ok {
-					hourly[path][day] = allhours
-				}
-			}
-		}
-	}
-
-	// Fill in blank hours.
-	for path, days := range hourly {
-		for dayk, day := range days {
-			if len(day) == 24 {
-				continue
-			}
-
-			newday := make([][]int, 24)
-		outer:
-			for i, hour := range allhours {
-				for _, h := range day {
-					if h[0] == hour[0] {
-						newday[i] = h
-						continue outer
-					}
-				}
-				newday[i] = hour
-			}
-
-			hourly[path][dayk] = newday
-		}
-	}
-
-	return hourly
 }