--- a/go.mod
+++ b/go.mod
@@ -4,18 +4,17 @@ go 1.12
 
 require (
 	github.com/arp242/geoip2-golang v1.4.0
-	github.com/fsnotify/fsnotify v1.4.7 // indirect
 	github.com/go-chi/chi v4.0.2+incompatible
 	github.com/jinzhu/now v1.1.1
 	github.com/jmoiron/sqlx v1.2.0
 	github.com/lib/pq v1.2.0
-	github.com/mattn/go-sqlite3 v1.11.0
+	github.com/mattn/go-sqlite3 v2.0.2+incompatible
 	github.com/mssola/user_agent v0.5.0
 	github.com/pkg/errors v0.8.1
 	github.com/teamwork/guru v1.0.0
 	github.com/teamwork/reload v1.2.1
 	zgo.at/utils v1.2.0
-	zgo.at/zdb v0.0.0-20191111044120-fca4c0a247fd
+	zgo.at/zdb v0.0.0-20191217115710-4a6a7376dd60
 	zgo.at/zhttp v0.0.0-20191211182637-d99b56ea9f13
 	zgo.at/zlog v1.0.6
 	zgo.at/zpack v1.0.0