Re-building everyone was a good prototype, but doesn't scale. It's
already taking ~800ms on production servers.

Before:

    20:38:13 cron: INFO: persisted 1 hits {memstore=7ms stats=1142ms}

Now:

    20:36:34 cron: INFO: persisted 1 hits {memstore=7ms stats=16ms}